Error tracking, backup before converting math classes to use SIMD.

+++ OpenGL 3D Game/src/engine/platform/opengl/buffer/GLMeshVBO.cpp
@@ -36,11 +36,11 @@ void GLMeshVBO::Store(const Mesh& mesh)
 	glEnableVertexAttribArray(0);
 	glEnableVertexAttribArray(1);
 	glEnableVertexAttribArray(2);
-
-	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(float) * Vertex::GetNumberOfFloats(), (void*) (sizeof(float) * 0));
-	glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, sizeof(float) * Vertex::GetNumberOfFloats(), (void*) (sizeof(float) * 3));
-	glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, sizeof(float) * Vertex::GetNumberOfFloats(), (void*) (sizeof(float) * 5));
-	glVertexAttribPointer(3, 3, GL_FLOAT, GL_FALSE, sizeof(float) * Vertex::GetNumberOfFloats(), (void*) (sizeof(float) * 8));
+	
+	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, int(sizeof(float) * Vertex::GetNumberOfFloats()), (void*) (sizeof(float) * 0));
+	glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, int(sizeof(float) * Vertex::GetNumberOfFloats()), (void*) (sizeof(float) * 3));
+	glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, int(sizeof(float) * Vertex::GetNumberOfFloats()), (void*) (sizeof(float) * 5));
+	glVertexAttribPointer(3, 3, GL_FLOAT, GL_FALSE, int(sizeof(float) * Vertex::GetNumberOfFloats()), (void*) (sizeof(float) * 8));
 
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
 	glBindBuffer(GL_ARRAY_BUFFER, 0);
